+++ Source/IGListSectionController.m
@@ -86,4 +86,8 @@ void IGListSectionControllerPopThread(void) {
 
 - (void)didDeselectItemAtIndex:(NSInteger)index {}
 
+- (void)didHighlightItemAtIndex:(NSInteger)index {}
+
+- (void)didUnhighlightItemAtIndex:(NSInteger)index {}
+
 @end